Player walking and player/camera rotation speed
The player can now walk around on the surface of planets. The player's rotation of having their feet move towards the "down" direction now scales with the player's speed for a smoother transition. There is currently a bug where the player/camera make sharp jumps when on the surface of the planet.