Custom player sprite, checkpoint sprite, slow fall
Added one custom sprite of the player in a blue space suit. Added 11 custom sprites for the checkpoint and its animation. Holding space now slows time. Keycodes required for walking and slowing time are now specified at the top of the player and cameracontroller scripts. The player now takes longer to slow down on a low-friction planet, and now comes to a complete stop.