Added pause functionality and improved the player's rotation on a planet
Esc now pauses the game, with a pause menu to be added soon. The amount of time a player spends on their current planet is now tracked. The player's rotation snaps to the ground only at very small angles. If the player spends a long amount of time on a planet, their rotation amount scales with the amount of time they have spent. Overall, player rotation toward the ground is now extremely smooth.