Gravity Gun v12
Time now stops for the player while they are rotating towards a planet. When the player clicks a planet, they will stop all momentum, rotate towards their destination planet, and then resume their momentum. If the player collides with a planet during their fall, they will stop momentum, rotate towards the planet they are on, and then resume momentum once again, now on the surface of the new planet. During the rotation process, the screen is tinted purple as a visual cue. The player can now click on a planet while falling, and they will rotate towards the new choice and begin falling towards it.