Gravity Gun v15
Updated the clicking/tapping detection to be able to handle multiple objects with tapping functions. Specifically, when the player taps something, a list is made of all of the possible objects they could have tapped, and the one with the highest priority is interacted with. Also, most of the variables and functions have been commented properly and sorted. Removed the checkpoint script and prefab.