Gravity Gun v16
Added the basics of the NPC dialogue system. The player can approach an NPC, and once within range, a thought bubble pops up above the NPC. Clicking the NPC or the thought bubble shows a dialogue box, which can be tapped to cycle through several lines of text. There are no unique visual assets for the NPCs and the dialogue system yet.