Gravity Gun v17
Improved the visuals for the Dialogue system. Added a function that splits dialogue text into single-line segments by inserting a new line character in between some words, so that the text fits within the horizontal confines of the text box. The text now displays gradually, adding one character per frame until all of the text is displayed. The player can tap the dialogue box to instantly display all of the text in the current sentence. Changed the touch detection script in cameracontroller to reference "Ignore Touch" and "Untappable" tags, rather than having to manually add those elements to a large list. The npcdialogue script now rotates the NPC toward whatever planet they are on when the scene starts.