The player can now hit obstacles mid-fall
They will either hit the side of the obstacle and slide down it, or hit the top and walk along the top of it. When the player is sliding down the side of the obstacle and lands on the planet, their movement continues as normal. When the player is walking along the top of an obstacle and loses contact with the obstacle, they begin an arching trajectory towards the planet as normal.