Modified getPlayerEdgePositions
the function getPlayerEdgePositions has been replaced with getPlayerSquareEdgePositions and getPlayerCircleEdgePositions. The square one behaves exactly the same as before, where it gets the 8 points in a square around the player. The circle one now gets the 8 points in a circle around the player. The two functions are nearly identical, except the circle one has the diagonal edge points on the circle's surface. The circle edge function is used to check for collisions with objects mid-fall. The square one is used to check if the player is still walking on top of on obstacle.