Obstacle fixes
At the start of the game, all obstacles are rotated towards their planets, in the same way as NPCDialogue objects are. When the player collides with an obstacle while leaving a planet, the player will no longer pass through it. Added more edge detection spots in GetPlayerCircleEdgePositions. The player is now prevented from clicking another planet if they are too close to an obstacle. That distance is currently defined as 1.2x the player's radius.