Updated FindDistanceToLine
FindDistanceToLine in the Player script now includes code that restricts the "closest found point" to be within the bounds of the line segment as described by the provided endpoints. In common english, now it will return the distance to a point that is definitely inside or on the endpoints of the provided line.