Added the start of a quest system
Removed some unused files. Changed NPCDialogue to NPC. Added the RotateTowardsPlanet script, which is to be temporarily added to any gameobject to rotate it towards its parent planet. After finding the right rotation coordinates, the script should be removed. Added the Quest and QuestManager scripts. The QuestManager creates several Quest objects, and each Quest contains data on what the player has to do to complete it, if the quest is active, and its prerequisites. Added the "Talk to NPC" quest requirement, which provides a specified NPC with specific quest dialogue. After finishing the dialogue, the quest is complete. NPCs also have non-quest dialogue, which is used by the DialogueManager system if there is no quest relating to the NPC.