Added more player animations
When the player taps a planet, the player switches to an animation that only includes one of the player sprite's arms. An image of the player's hand holding the gravity gun becomes visible and rotates to point at the desitnation planet, all in slowed-time. Then, a purple beam travels from the player to the planet, and then the normal rotation-in-stopped-time functionality continues as normal.  Added a purple "glow" sprite as a child of the Planet object, which is made visible once the gravity gun beam hits the planet.